refactor(sam): RC7 — clean up stale red-phase test artifacts, promote memory_backend fixture
- Promote memory_backend fixture to conftest.py, removing local
  definitions from test_drafting_state.py and test_large_document_write.py
- Rename _make_task_yaml -> _make_task_def in test_large_document_write.py
  (name now matches the TaskDefinition model it returns)
- Tighten pytest.raises((TaskValidationError, ValueError)) to
  pytest.raises(TaskValidationError) — RC4 guarantees the single exception
- Remove stale "RED-PHASE" comment blocks from test_server.py
  (append_task and finalize are implemented and green)
- Clean up resulting unused imports (pytest, TYPE_CHECKING, task_config)
